Move getFiles() to lib.rs
